fix: Resolve AI comments, add isAlreadyExists check to container driver
