+++ backend/src/v2/driver/root_dag.go
@@ -18,7 +18,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"strings"
 
 	"github.com/golang/glog"
 	"github.com/kubeflow/pipelines/api/v2alpha1/go/pipelinespec"
@@ -26,8 +25,6 @@ import (
 	"github.com/kubeflow/pipelines/backend/src/v2/config"
 	"github.com/kubeflow/pipelines/backend/src/v2/metadata"
 	"github.com/kubeflow/pipelines/backend/src/v2/objectstore"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 	"k8s.io/client-go/kubernetes"
 )
 
@@ -135,8 +132,8 @@ func RootDAG(ctx context.Context, opts Options, mlmd *metadata.Client) (executio
 		// The deterministic name "run/<runId>" causes a duplicate key error.
 		// Handle it here by looking up the existing record instead of failing.
 		if isAlreadyExistsErr(err) {
-			glog.Infof("Execution %q already exists (likely a retry), looking up existing execution", ecfg.Name)
-			existing, lookupErr := mlmd.GetExecutionsByTypeAndName(ctx, string(metadata.DagExecutionTypeName), ecfg.Name)
+			glog.Infof("Execution %q already exists, looking up existing execution", ecfg.Name)
+			existing, lookupErr := mlmd.GetExecutionByTypeAndName(ctx, string(metadata.DagExecutionTypeName), ecfg.Name)
 			if lookupErr != nil {
 				return nil, fmt.Errorf("failed to lookup existing execution: %w", lookupErr)
 			}
@@ -154,16 +151,4 @@ func RootDAG(ctx context.Context, opts Options, mlmd *metadata.Client) (executio
 	// No need to return ExecutorInput, because tasks in the DAG will resolve
 	// needed info from MLMD.
 	return &Execution{ID: exec.GetID()}, nil
-}
-
-// isAlreadyExistsErr checks whether the error is a gRPC AlreadyExists error, or contains
-// the "AlreadyExists" string in the error message.
-func isAlreadyExistsErr(err error) bool {
-	if s, ok := status.FromError(err); ok && s.Code() == codes.AlreadyExists {
-		return true
-	}
-
-	// MLMD sometimes wraps the ALreadyExists in an internal error, so also check the error message
-	// string as a fallback
-	return strings.Contains(err.Error(), "AlreadyExists") || strings.Contains(err.Error(), "Duplicate entry")
 }
