+++ backend/src/v2/driver/util_test.go
@@ -15,11 +15,14 @@
 package driver
 
 import (
+	"fmt"
 	"testing"
 
 	"github.com/kubeflow/pipelines/api/v2alpha1/go/pipelinespec"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 	structpb "google.golang.org/protobuf/types/known/structpb"
 )
 
@@ -54,6 +57,46 @@ func Test_isInputParameterChannel(t *testing.T) {
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
 			assert.Equal(t, isInputParameterChannel(test.input), test.isValid)
+		})
+	}
+}
+
+func Test_isAlreadyExistsErr(t *testing.T) {
+	tests := []struct {
+		name string
+		err  error
+		want bool
+	}{
+		{
+			name: "grpc already exists",
+			err:  status.Error(codes.AlreadyExists, "duplicate execution"),
+			want: true,
+		},
+		{
+			name: "wrapped grpc already exists",
+			err:  fmt.Errorf("wrapped: %w", status.Error(codes.AlreadyExists, "duplicate execution")),
+			want: true,
+		},
+		{
+			name: "already exists in message",
+			err:  fmt.Errorf("rpc error: code = Internal desc = AlreadyExists: execution exists"),
+			want: true,
+		},
+		{
+			name: "duplicate entry in message",
+			err:  fmt.Errorf("sql failure: Duplicate entry 'run/abc'"),
+			want: true,
+		},
+		{
+			name: "other error",
+			err:  status.Error(codes.Internal, "some other failure"),
+			want: false,
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			assert.Equal(t, test.want, isAlreadyExistsErr(test.err))
 		})
 	}
 }
